chore: extract `EligibilityCheckType` enum to a dedicated module
Moved `EligibilityCheckType` from `core.dtos.chat.rule` to `core.enums.rule` for improved modularity and better code organization. Updated all related imports accordingly.
