+++ backend/core/dtos/chat/rule/sticker.py
@@ -3,7 +3,8 @@ from typing import Self
 from pydantic import BaseModel, computed_field
 
 from core.constants import PROMOTE_STICKER_COLLECTION_TEMPLATE
-from core.dtos.chat.rule import ChatEligibilityRuleDTO, EligibilityCheckType
+from core.dtos.chat.rule import ChatEligibilityRuleDTO
+from core.enums.rule import EligibilityCheckType
 from core.dtos.chat.rule.internal import EligibilitySummaryStickerCollectionInternalDTO
 from core.dtos.sticker import MinimalStickerCollectionDTO, MinimalStickerCharacterDTO
 from core.models.rule import TelegramChatStickerCollection
